Move design_matrix_panel to where it is used
