+++ src/ert/gui/simulation/_design_matrix_panel.py
